+++ geometry/generate_random_points.c
@@ -16,7 +16,7 @@
 
 #include "geometry/generate_random_points.h"
 
-void generate_random_points(long (*rng)(), sp_func_t* density, bbox_t* bounding_box, int num_points, point_t* points)
+void generate_random_points(int (*rng)(), sp_func_t* density, bbox_t* bounding_box, int num_points, point_t* points)
 {
   ASSERT(density != NULL);
   ASSERT(bounding_box != NULL);
